Make default editor neovim and other improvements
If neovim is installed, it will effectively replace vim. Fix opencv library path issue. Add alias to quit and clear quicker in the terminal. Add kubectl and minikube zsh plugins. After restart of computer /tmp/whereami is empty and we get a warning. This has been fixed to first check if the file exists.