Add scratchpad terminal for use in i3wm
By using the i3 scratchpad feature we can create a scratchpad terminal. The
terminal is spawned and sent to the scratchpad. When we want to use it
we move it to our current workspace (and we have rules for its size and
position). When we no longer want it we can send it back to the scratchpad.
Scratchpad is not accessible like other i3 workspaces.