Enable termdebug and add terminal mode mappings
Use jk and kj as alternative for escape in insert ,command and
terminal mode. These keys keep you on the home row and are inverses of
each other in all the modes just mentioned.

The Terminal debugging plugin can be used to debug a program with gdb and view
the source code in a Vim window.  Since this is completely contained inside
Vim this also works remotely over an ssh connection.

Add mappings to enter normal mode from terminal mode. Enable and
configure termdebug window size.